+++ core/client/app/controllers/modals/delete-user.js
@@ -30,7 +30,7 @@ export default Ember.Controller.extend({
 
             user.destroyRecord().then(function () {
                 self.store.unloadAll('post');
-                self.transitionToRoute('settings.users');
+                self.transitionToRoute('team');
                 self.get('notifications').showSuccess('The user has been deleted.', {delayed: true});
             }, function () {
                 self.get('notifications').showError('The user could not be deleted. Please try again.');
