+++ core/client/app/mixins/current-user-settings.js
@@ -5,7 +5,7 @@ var CurrentUserSettings = Ember.Mixin.create({
 
         return function (user) {
             if (user.get('isAuthor')) {
-                return self.transitionTo('settings.users.user', user);
+                return self.transitionTo('team.user', user);
             }
 
             return user;
@@ -17,7 +17,7 @@ var CurrentUserSettings = Ember.Mixin.create({
 
         return function (user) {
             if (user.get('isEditor')) {
-                return self.transitionTo('settings.users');
+                return self.transitionTo('team');
             }
 
             return user;
