+++ core/client/app/router.js
@@ -40,17 +40,15 @@ Router.map(function () {
         this.route('edit', {path: ':post_id'});
     });
 
-    this.route('settings.general', {path: '/settings/general'});
-    this.route('settings.users', {path: '/settings/users'}, function () {
+    this.route('team', {path: '/team'}, function () {
         this.route('user', {path: ':slug'});
     });
+
+    this.route('settings.general', {path: '/settings/general'});
     this.route('settings.tags', {path: '/settings/tags'});
     this.route('settings.labs', {path: '/settings/labs'});
     this.route('settings.code-injection', {path: '/settings/code-injection'});
     this.route('settings.navigation', {path: '/settings/navigation'});
-
-    // Redirect legacy content to posts
-    this.route('content');
 
     this.route('error404', {path: '/*path'});
 });
