--- core/client/app/routes/content.js
@@ -1,8 +1,0 @@
-import Ember from 'ember';
-var ContentRoute = Ember.Route.extend({
-    beforeModel: function () {
-        this.transitionTo('posts');
-    }
-});
-
-export default ContentRoute;
