--- core/client/app/routes/debug.js
@@ -1,8 +1,0 @@
-import Ember from 'ember';
-var DebugRoute = Ember.Route.extend({
-    beforeModel: function () {
-        this.transitionTo('settings.labs');
-    }
-});
-
-export default DebugRoute;
