+++ core/client/app/templates/team/user.hbs
@@ -1,6 +1,6 @@
 <header class="view-header">
     {{#gh-view-title openMobileMenu="openMobileMenu"}}
-        {{#link-to "settings.users"}}Team{{/link-to}} <i class="icon-arrow-right"></i> {{user.name}}
+        {{#link-to "team"}}Team{{/link-to}} <i class="icon-arrow-right"></i> {{user.name}}
     {{/gh-view-title}}
     <section class="view-actions">
         {{#if view.userActionsAreVisible}}
