--- core/client/app/views/settings/users/index.js
@@ -1,5 +1,0 @@
-import BaseView from 'ghost/views/settings/content-base';
-
-var SettingsUserIndexView = BaseView.extend();
-
-export default SettingsUserIndexView;
