+++ core/client/app/views/team/index.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+var TeamUserIndexView = Ember.View.extend({
+    tagName: 'section',
+    classNames: ['gh-view']
+});
+
+export default TeamUserIndexView;
